+++ frontend/appflowy_flutter/lib/startup/tasks/app_widget.dart
@@ -1,7 +1,10 @@
 import 'dart:io';
 
+import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
+
 import 'package:appflowy/mobile/application/mobile_router.dart';
-import 'package:appflowy/plugins/document/presentation/more/cubit/document_appearance_cubit.dart';
+import 'package:appflowy/plugins/document/application/document_appearance_cubit.dart';
 import 'package:appflowy/startup/startup.dart';
 import 'package:appflowy/user/application/reminder/reminder_bloc.dart';
 import 'package:appflowy/user/application/user_settings_service.dart';
@@ -19,8 +22,6 @@ import 'package:appflowy_editor/appflowy_editor.dart' hide Log;
 import 'package:easy_localization/easy_localization.dart';
 import 'package:flowy_infra/theme.dart';
 import 'package:flowy_infra_ui/flowy_infra_ui.dart';
-import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:go_router/go_router.dart';
 
@@ -154,8 +155,7 @@ class _ApplicationWidgetState extends State<ApplicationWidget> {
           value: getIt<ReminderBloc>()..add(const ReminderEvent.started()),
         ),
         BlocProvider<ViewInfoBloc>(
-          create: (_) =>
-              getIt<ViewInfoBloc>()..add(const ViewInfoEvent.started()),
+          create: (_) => getIt<ViewInfoBloc>(),
         ),
       ],
       child: BlocListener<NotificationActionBloc, NotificationActionState>(
