+++ frontend/appflowy_flutter/lib/workspace/presentation/widgets/more_view_actions/more_view_actions.dart
@@ -4,9 +4,11 @@ import 'package:appflowy/generated/flowy_svgs.g.dart';
 import 'package:appflowy/generated/locale_keys.g.dart';
 import 'package:appflowy/startup/startup.dart';
 import 'package:appflowy/workspace/application/settings/appearance/appearance_cubit.dart';
-import 'package:appflowy/workspace/application/settings/date_time/date_format_ext.dart';
 import 'package:appflowy/workspace/application/view_info/view_info_bloc.dart';
-import 'package:appflowy/workspace/presentation/widgets/view_actions_top_bar/view_actions.dart';
+import 'package:appflowy/workspace/presentation/widgets/more_view_actions/widgets/delete_view_action.dart';
+import 'package:appflowy/workspace/presentation/widgets/more_view_actions/widgets/duplicate_view_action.dart';
+import 'package:appflowy/workspace/presentation/widgets/more_view_actions/widgets/font_size_action.dart';
+import 'package:appflowy/workspace/presentation/widgets/more_view_actions/widgets/view_meta_info.dart';
 import 'package:appflowy_backend/protobuf/flowy-folder/view.pb.dart';
 import 'package:appflowy_backend/protobuf/flowy-user/date_time.pbenum.dart';
 import 'package:appflowy_popover/appflowy_popover.dart';
@@ -16,16 +18,24 @@ import 'package:flowy_infra_ui/style_widget/hover.dart';
 import 'package:flowy_infra_ui/widget/flowy_tooltip.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
-class DatabaseMoreAction extends StatefulWidget {
-  const DatabaseMoreAction({super.key, required this.view});
+class MoreViewActions extends StatefulWidget {
+  const MoreViewActions({
+    super.key,
+    required this.view,
+    this.isDocument = true,
+  });
 
+  /// The view to show the actions for.
   final ViewPB view;
 
+  /// If false the view is a Database, otherwise it is a Document.
+  final bool isDocument;
+
   @override
-  State<DatabaseMoreAction> createState() => _DatabaseMoreActionState();
+  State<MoreViewActions> createState() => _MoreViewActionsState();
 }
 
-class _DatabaseMoreActionState extends State<DatabaseMoreAction> {
+class _MoreViewActionsState extends State<MoreViewActions> {
   late final UserDateFormatPB dateFormat;
   final popoverMutex = PopoverMutex();
 
@@ -53,10 +63,16 @@ class _DatabaseMoreActionState extends State<DatabaseMoreAction> {
             offset: const Offset(0, 30),
             popupBuilder: (_) {
               final actions = [
+                if (widget.isDocument) ...[
+                  const FontSizeAction(),
+                  const Divider(height: 4),
+                ],
                 DuplicateViewAction(view: widget.view, mutex: popoverMutex),
                 DeleteViewAction(view: widget.view, mutex: popoverMutex),
-                _MoreActionFooter(
+                const Divider(height: 4),
+                ViewMetaInfo(
                   dateFormat: dateFormat,
+                  documentCounters: state.documentCounters,
                   createdAt: state.createdAt,
                 ),
               ];
@@ -73,57 +89,24 @@ class _DatabaseMoreActionState extends State<DatabaseMoreAction> {
             child: FlowyTooltip(
               message: LocaleKeys.moreAction_moreOptions.tr(),
               child: FlowyHover(
-                child: Padding(
+                style: HoverStyle(
+                  foregroundColorOnHover:
+                      Theme.of(context).colorScheme.onPrimary,
+                ),
+                builder: (context, isHovering) => Padding(
                   padding: const EdgeInsets.all(6),
                   child: FlowySvg(
                     FlowySvgs.details_s,
                     size: const Size(18, 18),
-                    color: Theme.of(context).iconTheme.color,
+                    color: isHovering
+                        ? Theme.of(context).colorScheme.onPrimary
+                        : Theme.of(context).iconTheme.color,
                   ),
                 ),
               ),
             ),
           );
         },
-      ),
-    );
-  }
-}
-
-class _MoreActionFooter extends StatelessWidget {
-  const _MoreActionFooter({
-    required this.dateFormat,
-    this.createdAt,
-  });
-
-  final UserDateFormatPB dateFormat;
-  final DateTime? createdAt;
-
-  @override
-  Widget build(BuildContext context) {
-    if (createdAt == null) {
-      return const SizedBox.shrink();
-    }
-
-    return Padding(
-      padding: const EdgeInsets.symmetric(horizontal: 6),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          if (createdAt != null) ...[
-            const Divider(height: 2),
-            const VSpace(2),
-            FlowyText(
-              LocaleKeys.moreAction_createdAt.tr(
-                args: [
-                  dateFormat.formatDate(createdAt!, false),
-                ],
-              ),
-              color: Theme.of(context).hintColor,
-              fontSize: 10,
-            ),
-          ],
-        ],
       ),
     );
   }
